history module refactoring

+++ src/history/history.service.ts
@@ -1,8 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose'
 import { RecurringBill } from 'src/recurring-bills/interfaces/recurring-bills.interface';
 import { HistoryItem, LineChartData } from './interfaces/history-item.interface';
+import { nextMonthDate } from 'src/utils/general_utils';
+import mongoose from 'mongoose';
 
 
 @Injectable()
@@ -12,6 +14,15 @@ export class HistoryService {
         @InjectModel('RecurringBill') private readonly recurringBillModel:Model<RecurringBill>
     ) {}
 
+    async getAllFromBill(billId: string): Promise<HistoryItem[]> {
+        const isValidId =  mongoose.isValidObjectId(billId)
+        if(!isValidId) {
+            throw new BadRequestException('Invalid ID: please enter a valid ID')
+        }
+        const items = await this.historyItemModel.find({recurringBillId: billId})
+        return items
+    }
+        
     async getAll(): Promise<HistoryItem[]> {
         return await this.historyItemModel.find()
     }
@@ -41,16 +52,37 @@ export class HistoryService {
     }
 
     async getOne(id: string): Promise<HistoryItem> {
+        const isValidId =  mongoose.isValidObjectId(id)
+        if(!isValidId) {
+            throw new BadRequestException('Invalid ID: please enter a valid ID')
+        }
         return await this.historyItemModel.findOne({ _id: id});
     }
 
     async create(newItem: HistoryItem): Promise<HistoryItem> {
+        const isValidId =  mongoose.isValidObjectId(newItem.recurringBillId)
+        if(!isValidId) {
+            throw new BadRequestException('Invalid ID: please enter a valid ID')
+        }
+        
+        const nextExpirationDate = nextMonthDate(newItem.expirationDate)
 
-        const newHistoryItem = new this.historyItemModel(newItem)
+        const updatedBill = { nextExpirationDate }
+        
+        const newBill = await this.recurringBillModel.findByIdAndUpdate(newItem.recurringBillId, updatedBill,{new: true})
+
+        const formatedHistoryItem = Object.assign(newItem, {title: newBill.title})
+
+        const newHistoryItem = new this.historyItemModel(formatedHistoryItem)
+        
         return await newHistoryItem.save()
     }
 
     async delete(id: string): Promise<HistoryItem> {
+        const isValidId =  mongoose.isValidObjectId(id)
+        if(!isValidId) {
+            throw new BadRequestException('Invalid ID: please enter a valid ID')
+        }
         return await this.historyItemModel.findByIdAndRemove(id)
     }
 }
